BioTIMEr: Tools to Use and Explore the 'BioTIME' Database

The BioTIMEr package is developed in GitHub (https://github.com/bioTIMEHub/BioTIMEr). To see the preferable citation of the package, type citation("BioTIMEr"). The associated vignette includes an introduction to the functions, how to apply them as well as some hints on how to quantify and visualise temporal biodiversity change for a given BioTIME dataset.

See also

Author

Maintainer: Alban Sagouis alban.sagouis@idiv.de (ORCID)

Authors:

Faye Moyes (ORCID)

Inês S. Martins (ORCID) [reviewer]

Other contributors:

Shane A. Blowes (ORCID) [contributor]

Viviana Brambilla (ORCID) [contributor]

Cher F. Y. Chow (ORCID) [contributor]

Ada Fontrodona-Eslava (ORCID) [contributor]

Laura Antão (ORCID) [contributor, reviewer]

Jonathan M. Chase (ORCID) [funder]

Maria Dornelas (ORCID) [funder, copyright holder]

Anne E. Magurran (ORCID) [funder]

European Research Council grant AdG BioTIME 250189 [funder]

European Research Council grant PoC BioCHANGE 727440 [funder]

European Research Council grant AdG MetaCHANGE 101098020 [funder]

The Leverhulme Centre for Anthropocene Biodiversity grant RC-2018-021 [funder]

German Centre for Integrative Biodiversity Research (iDiv) Halle-Jena-Leipzig (ROR) [funder]

Martin Luther University Halle-Wittenberg (ROR) [funder]

University of St Andrews (ROR) [funder]