BioTIMEr: Tools to Use and Explore the BioTIME Database

The BioTIMEr package is developed at GitHub (https://github.com/bioTIMEHub/BioTIMEr). To see the preferable citation of the package, type citation("BioTIMEr"). Changes in BioTIMEr documents can be read with [add code]. The associated vignette includes an introduction to the functions, how to apply them as well as some hints on how to quantify and visualise temporal biodiversity change for a given BioTIME dataset.

See also

Author

Maintainer: Alban Sagouis alban.sagouis@idiv.de (ORCID)

Authors:

Faye Moyes (ORCID)

Inês Santos Martins (ORCID)

Other contributors:

Shane A. Blowes (ORCID) [contributor]

Viviana Brambilla (ORCID) [contributor]

Cher F. Y. Chow (ORCID) [contributor]

Ada Fontrodona-Eslava (ORCID) [contributor]

Anne E. Magurran (ORCID) [funder]

Jonathan M. Chase (ORCID) [funder]

Maria Dornelas (ORCID) [funder, copyright holder]

European Research Council grant AdG BioTIME 250189 [funder]

European Research Council grant PoC BioCHANGE 727440 [funder]